split dataclasses into energy_demand.py and energy_source.py
