+++ generatorcore/agri2030/energy_demand.py
@@ -1,0 +1,719 @@
+# pyright: strict
+from dataclasses import dataclass
+from typing import Any
+
+from ..utils import div
+from ..agri2018.a18 import A18
+from ..inputs import Inputs
+
+
+@dataclass
+class CO2eChangeP:
+    CO2e_production_based: float = None  # type: ignore
+    CO2e_total: float = None  # type: ignore
+    CO2e_total_2021_estimated: float = None  # type: ignore
+    change_CO2e_pct: float = None  # type: ignore
+    change_CO2e_t: float = None  # type: ignore
+    cost_climate_saved: float = None  # type: ignore
+    cost_wage: float = None  # type: ignore
+    demand_emplo: float = None  # type: ignore
+    demand_emplo_new: float = None  # type: ignore
+    energy: float = None  # type: ignore
+    invest: float = None  # type: ignore
+    invest_pa: float = None  # type: ignore
+
+    @classmethod
+    def calc(
+        cls,
+        inputs: Inputs,
+        what: str,
+        a18: A18,
+        p_operation: Any,
+        CO2e_production_based: float,
+        CO2e_total: float,
+    ) -> "CO2eChangeP":
+
+        a18_CO2e_total = getattr(a18, what).CO2e_total
+
+        CO2e_total_2021_estimated = a18_CO2e_total * inputs.fact(
+            "Fact_M_CO2e_wo_lulucf_2021_vs_2018"
+        )
+
+        change_CO2e_t = CO2e_total - a18_CO2e_total
+        cost_climate_saved = (
+            (CO2e_total_2021_estimated - CO2e_total)
+            * inputs.entries.m_duration_neutral
+            * inputs.fact("Fact_M_cost_per_CO2e_2020")
+        )
+        change_CO2e_pct = div(change_CO2e_t, a18_CO2e_total)
+        demand_emplo = p_operation.demand_emplo
+
+        invest = p_operation.invest
+        invest_pa = invest / inputs.entries.m_duration_target
+
+        demand_emplo_new = p_operation.demand_emplo_new
+        energy = p_operation.energy
+        cost_wage = p_operation.cost_wage
+
+        return cls(
+            CO2e_production_based=CO2e_production_based,
+            CO2e_total=CO2e_total,
+            CO2e_total_2021_estimated=CO2e_total_2021_estimated,
+            change_CO2e_pct=change_CO2e_pct,
+            change_CO2e_t=change_CO2e_t,
+            cost_climate_saved=cost_climate_saved,
+            cost_wage=cost_wage,
+            demand_emplo=demand_emplo,
+            demand_emplo_new=demand_emplo_new,
+            energy=energy,
+            invest=invest,
+            invest_pa=invest_pa,
+        )
+
+
+@dataclass
+class CO2eChange:
+    CO2e_combustion_based: float = None  # type: ignore
+    CO2e_production_based: float = None  # type: ignore
+    CO2e_total: float = None  # type: ignore
+    CO2e_total_2021_estimated: float = None  # type: ignore
+    change_CO2e_pct: float = None  # type: ignore
+    change_CO2e_t: float = None  # type: ignore
+    cost_climate_saved: float = None  # type: ignore
+
+    @classmethod
+    def calc(
+        cls,
+        inputs: Inputs,
+        what: str,
+        a18: A18,
+        CO2e_combustion_based: float,
+        CO2e_production_based: float,
+    ) -> "CO2eChange":
+
+        if not what:
+            a18_CO2e_total = 0
+        else:
+            a18_CO2e_total = getattr(a18, what).CO2e_total
+
+        CO2e_total = CO2e_production_based + CO2e_combustion_based
+        change_CO2e_t = CO2e_total - a18_CO2e_total
+        change_CO2e_pct = div(change_CO2e_t, a18_CO2e_total)
+
+        CO2e_total_2021_estimated = a18_CO2e_total * inputs.fact(
+            "Fact_M_CO2e_wo_lulucf_2021_vs_2018"
+        )
+        cost_climate_saved = (
+            (CO2e_total_2021_estimated - CO2e_total)
+            * inputs.entries.m_duration_neutral
+            * inputs.fact("Fact_M_cost_per_CO2e_2020")
+        )
+
+        return cls(
+            CO2e_combustion_based=CO2e_combustion_based,
+            CO2e_production_based=CO2e_production_based,
+            CO2e_total=CO2e_total,
+            CO2e_total_2021_estimated=CO2e_total_2021_estimated,
+            change_CO2e_pct=change_CO2e_pct,
+            change_CO2e_t=change_CO2e_t,
+            cost_climate_saved=cost_climate_saved,
+        )
+
+
+@dataclass
+class CO2eChangeFermentationOrManure(CO2eChange):
+    CO2e_combustion_based: float = None  # type: ignore
+    CO2e_production_based: float = None  # type: ignore
+    CO2e_production_based_per_t: float = None  # type: ignore
+    CO2e_total: float = None  # type: ignore
+    CO2e_total_2021_estimated: float = None  # type: ignore
+    amount: float = None  # type: ignore
+    change_CO2e_pct: float = None  # type: ignore
+    change_CO2e_t: float = None  # type: ignore
+    cost_climate_saved: float = None  # type: ignore
+    demand_change: float = None  # type: ignore
+
+    @classmethod
+    def calc_fermen(
+        cls, inputs: Inputs, what: str, ass_demand_change: str, a18: A18
+    ) -> "CO2eChangeFermentationOrManure":
+        demand_change = inputs.ass(ass_demand_change)
+        amount = getattr(a18, what).amount * (1 + demand_change)
+
+        CO2e_production_based_per_t = getattr(a18, what).CO2e_production_based_per_t
+        CO2e_production_based = amount * CO2e_production_based_per_t
+
+        parent = super().calc(inputs, what, a18, 0, CO2e_production_based)
+
+        return cls(
+            CO2e_combustion_based=parent.CO2e_combustion_based,
+            CO2e_production_based=parent.CO2e_production_based,
+            CO2e_production_based_per_t=CO2e_production_based_per_t,
+            CO2e_total=parent.CO2e_total,
+            CO2e_total_2021_estimated=parent.CO2e_total_2021_estimated,
+            amount=amount,
+            change_CO2e_pct=parent.change_CO2e_pct,
+            change_CO2e_t=parent.change_CO2e_t,
+            cost_climate_saved=parent.cost_climate_saved,
+            demand_change=demand_change,
+        )
+
+    @classmethod
+    def calc_manure(
+        cls, inputs: Inputs, what: str, a18: A18, amount: float
+    ) -> "CO2eChangeFermentationOrManure":
+        demand_change = inputs.ass("Ass_A_P_manure_ratio_CO2e_to_amount_change")
+
+        CO2e_production_based_per_t = getattr(a18, what).CO2e_production_based_per_t * (
+            1 + demand_change
+        )
+        CO2e_production_based = amount * CO2e_production_based_per_t
+
+        parent = super().calc(inputs, what, a18, 0, CO2e_production_based)
+
+        return cls(
+            CO2e_combustion_based=parent.CO2e_combustion_based,
+            CO2e_production_based=parent.CO2e_production_based,
+            CO2e_production_based_per_t=CO2e_production_based_per_t,
+            CO2e_total=parent.CO2e_total,
+            CO2e_total_2021_estimated=parent.CO2e_total_2021_estimated,
+            amount=amount,
+            change_CO2e_pct=parent.change_CO2e_pct,
+            change_CO2e_t=parent.change_CO2e_t,
+            cost_climate_saved=parent.cost_climate_saved,
+            demand_change=demand_change,
+        )
+
+
+@dataclass
+class CO2eChangeSoil(CO2eChange):
+    CO2e_combustion_based: float = None  # type: ignore
+    CO2e_production_based: float = None  # type: ignore
+    CO2e_production_based_per_t: float = None  # type: ignore
+    CO2e_total: float = None  # type: ignore
+    CO2e_total_2021_estimated: float = None  # type: ignore
+    area_ha: float = None  # type: ignore
+    area_ha_change: float = None  # type: ignore
+    change_CO2e_pct: float = None  # type: ignore
+    change_CO2e_t: float = None  # type: ignore
+    cost_climate_saved: float = None  # type: ignore
+    demand_change: float = None  # type: ignore
+
+    @classmethod
+    def calc_soil_grazing(
+        cls,
+        inputs: Inputs,
+        what: str,
+        a18: A18,
+        area_ha: float,
+        p_fermen_dairycow: Any,
+        p_fermen_nondairy: Any,
+        p_fermen_oanimal: Any,
+    ) -> "CO2eChangeSoil":
+
+        CO2e_production_based_per_t = div(
+            getattr(a18, what).CO2e_production_based_per_t
+            * (
+                p_fermen_dairycow.amount
+                + p_fermen_nondairy.amount
+                + p_fermen_oanimal.amount
+            ),
+            a18.p_fermen_dairycow.amount
+            + a18.p_fermen_nondairy.amount
+            + a18.p_fermen_oanimal.amount,
+        )
+
+        demand_change = (
+            div(
+                CO2e_production_based_per_t,
+                getattr(a18, what).CO2e_production_based_per_t,
+            )
+            - 1
+        )
+
+        CO2e_production_based = area_ha * CO2e_production_based_per_t
+
+        area_ha_change = -(getattr(a18, what).area_ha - area_ha)
+
+        parent = super().calc(inputs, what, a18, 0, CO2e_production_based)
+
+        return cls(
+            CO2e_combustion_based=parent.CO2e_combustion_based,
+            CO2e_production_based=parent.CO2e_production_based,
+            CO2e_production_based_per_t=CO2e_production_based_per_t,
+            CO2e_total=parent.CO2e_total,
+            CO2e_total_2021_estimated=parent.CO2e_total_2021_estimated,
+            area_ha=area_ha,
+            area_ha_change=area_ha_change,
+            change_CO2e_pct=parent.change_CO2e_pct,
+            change_CO2e_t=parent.change_CO2e_t,
+            cost_climate_saved=parent.cost_climate_saved,
+            demand_change=demand_change,
+        )
+
+    @classmethod
+    def calc_soil_residue(
+        cls, inputs: Inputs, what: str, a18: A18, area_ha: float
+    ) -> "CO2eChangeSoil":
+        CO2e_production_based_per_t = getattr(a18, what).CO2e_production_based_per_t
+        demand_change = (
+            div(
+                CO2e_production_based_per_t,
+                getattr(a18, what).CO2e_production_based_per_t,
+            )
+            - 1
+        )
+
+        CO2e_production_based = area_ha * CO2e_production_based_per_t
+
+        area_ha_change = -(getattr(a18, what).area_ha - area_ha)
+
+        parent = super().calc(inputs, what, a18, 0, CO2e_production_based)
+
+        return cls(
+            CO2e_combustion_based=parent.CO2e_combustion_based,
+            CO2e_production_based=parent.CO2e_production_based,
+            CO2e_production_based_per_t=CO2e_production_based_per_t,
+            CO2e_total=parent.CO2e_total,
+            CO2e_total_2021_estimated=parent.CO2e_total_2021_estimated,
+            area_ha=area_ha,
+            area_ha_change=area_ha_change,
+            change_CO2e_pct=parent.change_CO2e_pct,
+            change_CO2e_t=parent.change_CO2e_t,
+            cost_climate_saved=parent.cost_climate_saved,
+            demand_change=demand_change,
+        )
+
+    @classmethod
+    def calc_soil(
+        cls, inputs: Inputs, what: str, a18: A18, area_ha: float
+    ) -> "CO2eChangeSoil":
+        demand_change = inputs.ass("Ass_A_P_soil_N_application_2030_change")
+        CO2e_production_based_per_t = getattr(a18, what).CO2e_production_based_per_t * (
+            1 + demand_change
+        )
+
+        CO2e_production_based = area_ha * CO2e_production_based_per_t
+
+        area_ha_change = -(getattr(a18, what).area_ha - area_ha)
+
+        parent = super().calc(inputs, what, a18, 0, CO2e_production_based)
+
+        return cls(
+            CO2e_combustion_based=parent.CO2e_combustion_based,
+            CO2e_production_based=parent.CO2e_production_based,
+            CO2e_production_based_per_t=CO2e_production_based_per_t,
+            CO2e_total=parent.CO2e_total,
+            CO2e_total_2021_estimated=parent.CO2e_total_2021_estimated,
+            area_ha=area_ha,
+            area_ha_change=area_ha_change,
+            change_CO2e_pct=parent.change_CO2e_pct,
+            change_CO2e_t=parent.change_CO2e_t,
+            cost_climate_saved=parent.cost_climate_saved,
+            demand_change=demand_change,
+        )
+
+
+@dataclass
+class CO2eChangeOtherLiming(CO2eChange):
+    CO2e_combustion_based: float = None  # type: ignore
+    CO2e_production_based: float = None  # type: ignore
+    CO2e_total: float = None  # type: ignore
+    CO2e_total_2021_estimated: float = None  # type: ignore
+    change_CO2e_pct: float = None  # type: ignore
+    change_CO2e_t: float = None  # type: ignore
+    cost_climate_saved: float = None  # type: ignore
+    prod_volume: float = None  # type: ignore
+
+    @classmethod
+    def calc_other_liming(
+        cls,
+        inputs: Inputs,
+        what: str,
+        a18: A18,
+        prod_volume: float,
+        CO2e_production_based: float,
+    ) -> "CO2eChangeOtherLiming":
+
+        parent = super().calc(inputs, what, a18, 0, CO2e_production_based)
+
+        return cls(
+            CO2e_combustion_based=parent.CO2e_combustion_based,
+            CO2e_production_based=parent.CO2e_production_based,
+            CO2e_total=parent.CO2e_total,
+            CO2e_total_2021_estimated=parent.CO2e_total_2021_estimated,
+            change_CO2e_pct=parent.change_CO2e_pct,
+            change_CO2e_t=parent.change_CO2e_t,
+            cost_climate_saved=parent.cost_climate_saved,
+            prod_volume=prod_volume,
+        )
+
+
+@dataclass
+class CO2eChangeOther(CO2eChange):
+    CO2e_combustion_based: float = None  # type: ignore
+    CO2e_production_based: float = None  # type: ignore
+    CO2e_production_based_per_t: float = None  # type: ignore
+    CO2e_total: float = None  # type: ignore
+    CO2e_total_2021_estimated: float = None  # type: ignore
+    change_CO2e_pct: float = None  # type: ignore
+    change_CO2e_t: float = None  # type: ignore
+    cost_climate_saved: float = None  # type: ignore
+    demand_change: float = None  # type: ignore
+    prod_volume: float = None  # type: ignore
+
+    @classmethod
+    def calc_other(
+        cls,
+        inputs: Inputs,
+        what: str,
+        a18: A18,
+        ass_demand_change: str,
+        fact_production_based_per_t: str,
+    ) -> "CO2eChangeOther":
+
+        demand_change = inputs.ass(ass_demand_change)
+        prod_volume = getattr(a18, what).prod_volume * (1 + demand_change)
+
+        CO2e_production_based_per_t = inputs.fact(fact_production_based_per_t)
+        CO2e_production_based = prod_volume * CO2e_production_based_per_t
+
+        parent = super().calc(inputs, what, a18, 0, CO2e_production_based)
+
+        return cls(
+            CO2e_combustion_based=parent.CO2e_combustion_based,
+            CO2e_production_based=parent.CO2e_production_based,
+            CO2e_production_based_per_t=CO2e_production_based_per_t,
+            CO2e_total=parent.CO2e_total,
+            CO2e_total_2021_estimated=parent.CO2e_total_2021_estimated,
+            change_CO2e_pct=parent.change_CO2e_pct,
+            change_CO2e_t=parent.change_CO2e_t,
+            cost_climate_saved=parent.cost_climate_saved,
+            demand_change=demand_change,
+            prod_volume=prod_volume,
+        )
+
+
+@dataclass
+class CO2eChangePOperation:
+    change_energy_MWh: float = None  # type: ignore
+    change_energy_pct: float = None  # type: ignore
+    cost_wage: float = None  # type: ignore
+    demand_biomass: float = None  # type: ignore
+    demand_ediesel: float = None  # type: ignore
+    demand_electricity: float = None  # type: ignore
+    demand_emethan: float = None  # type: ignore
+    demand_emplo: float = None  # type: ignore
+    demand_emplo_new: float = None  # type: ignore
+    demand_epetrol: float = None  # type: ignore
+    demand_heatpump: float = None  # type: ignore
+    energy: float = None  # type: ignore
+    invest: float = None  # type: ignore
+    invest_pa: float = None  # type: ignore
+
+    @classmethod
+    def calc(
+        cls,
+        inputs: Inputs,
+        what: str,
+        a18: A18,
+        p_operation_vehicles: Any,
+        p_operation_heat: Any,
+        p_operation_elec_elcon: Any,
+        p_operation_elec_heatpump: Any,
+    ) -> "CO2eChangePOperation":
+        demand_epetrol = p_operation_vehicles.demand_epetrol
+        demand_ediesel = p_operation_vehicles.demand_ediesel
+        demand_heatpump = p_operation_heat.demand_heatpump
+        demand_emplo = p_operation_heat.demand_emplo
+        demand_emplo_new = p_operation_heat.demand_emplo_new
+        demand_biomass = p_operation_heat.demand_biomass
+        demand_electricity = (
+            p_operation_elec_elcon.demand_electricity
+            + p_operation_elec_heatpump.demand_electricity
+        )
+        demand_emethan = p_operation_heat.demand_emethan
+
+        energy = (
+            p_operation_heat.energy
+            + p_operation_elec_elcon.energy
+            + p_operation_elec_heatpump.energy
+            + p_operation_vehicles.energy
+        )
+
+        invest = p_operation_heat.invest
+        invest_pa = invest / inputs.entries.m_duration_target
+        cost_wage = p_operation_heat.cost_wage
+
+        change_energy_MWh = energy - getattr(a18, what).energy
+        change_energy_pct = div(change_energy_MWh, getattr(a18, what).energy)
+
+        return cls(
+            change_energy_MWh=change_energy_MWh,
+            change_energy_pct=change_energy_pct,
+            cost_wage=cost_wage,
+            demand_biomass=demand_biomass,
+            demand_ediesel=demand_ediesel,
+            demand_electricity=demand_electricity,
+            demand_emethan=demand_emethan,
+            demand_emplo=demand_emplo,
+            demand_emplo_new=demand_emplo_new,
+            demand_epetrol=demand_epetrol,
+            demand_heatpump=demand_heatpump,
+            energy=energy,
+            invest=invest,
+            invest_pa=invest_pa,
+        )
+
+
+@dataclass
+class CO2eChangePOperationHeat:
+    area_m2: float = None  # type: ignore
+    area_m2_nonrehab: float = None  # type: ignore
+    area_m2_rehab: float = None  # type: ignore
+    change_energy_MWh: float = None  # type: ignore
+    change_energy_pct: float = None  # type: ignore
+    cost_wage: float = None  # type: ignore
+    demand_biomass: float = None  # type: ignore
+    demand_ediesel: float = None  # type: ignore
+    demand_electricity: float = None  # type: ignore
+    demand_emethan: float = None  # type: ignore
+    demand_emplo: float = None  # type: ignore
+    demand_emplo_new: float = None  # type: ignore
+    demand_epetrol: float = None  # type: ignore
+    demand_heat_nonrehab: float = None  # type: ignore
+    demand_heat_rehab: float = None  # type: ignore
+    demand_heatpump: float = None  # type: ignore
+    emplo_existing: float = None  # type: ignore
+    energy: float = None  # type: ignore
+    fec_factor_averaged: float = None  # type: ignore
+    invest: float = None  # type: ignore
+    invest_pa: float = None  # type: ignore
+    invest_per_x: float = None  # type: ignore
+    pct_nonrehab: float = None  # type: ignore
+    pct_of_wage: float = None  # type: ignore
+    pct_rehab: float = None  # type: ignore
+    rate_rehab_pa: float = None  # type: ignore
+    ratio_wage_to_emplo: float = None  # type: ignore
+
+    @classmethod
+    def calc(
+        cls,
+        inputs: Inputs,
+        what: str,
+        a18: A18,
+    ) -> "CO2eChangePOperationHeat":
+        rate_rehab_pa = inputs.entries.r_rehab_rate_pa
+        pct_rehab = (
+            inputs.fact("Fact_B_P_ratio_renovated_to_not_renovated_2021")
+            + rate_rehab_pa * inputs.entries.m_duration_target
+        )
+        pct_nonrehab = 1 - pct_rehab
+
+        area_m2 = getattr(a18, what).area_m2
+        area_m2_rehab = pct_rehab * getattr(a18, what).area_m2
+        area_m2_nonrehab = pct_nonrehab * getattr(a18, what).area_m2
+
+        invest_per_x = inputs.fact("Fact_R_P_energetical_renovation_cost_business")
+        invest = (
+            area_m2_rehab
+            * (1 - inputs.fact("Fact_B_P_ratio_renovated_to_not_renovated_2021"))
+            * invest_per_x
+        )
+        invest_pa = invest / inputs.entries.m_duration_target
+
+        pct_of_wage = inputs.fact(
+            "Fact_B_P_renovations_ratio_wage_to_main_revenue_2017"
+        )
+        cost_wage = div(invest, inputs.entries.m_duration_target) * pct_of_wage
+
+        ratio_wage_to_emplo = inputs.fact(
+            "Fact_B_P_renovations_wage_per_person_per_year_2017"
+        )
+        emplo_existing = (
+            inputs.fact("Fact_B_P_renovation_emplo_2017")
+            * inputs.ass("Ass_B_D_renovation_emplo_pct_of_A")
+            * inputs.entries.m_population_com_2018
+            / inputs.entries.m_population_nat
+        )
+
+        demand_electricity = 0
+        demand_epetrol = 0
+        demand_ediesel = 0
+        demand_heat_rehab = area_m2_rehab * inputs.ass("Ass_B_D_ratio_fec_to_area_2050")
+        demand_heat_nonrehab = (
+            area_m2_nonrehab
+            * (
+                getattr(a18, what).factor_adapted_to_fec
+                - inputs.fact("Fact_B_P_ratio_renovated_to_not_renovated_2021")
+                * inputs.ass("Ass_B_D_ratio_fec_to_area_2050")
+            )
+            / (1 - inputs.fact("Fact_B_P_ratio_renovated_to_not_renovated_2021"))
+        )
+        demand_heatpump = demand_heat_rehab
+        demand_emplo = div(cost_wage, ratio_wage_to_emplo)
+        demand_emplo_new = max(0, demand_emplo - emplo_existing)
+
+        energy = demand_heat_nonrehab + demand_heat_rehab
+
+        demand_biomass = min(a18.s_biomass.energy, energy - demand_heatpump)
+        demand_emethan = energy - demand_biomass - demand_heatpump
+
+        fec_factor_averaged = div(energy, getattr(a18, what).area_m2)
+        change_energy_MWh = energy - getattr(a18, what).energy
+        change_energy_pct = div(change_energy_MWh, getattr(a18, what).energy)
+
+        return cls(
+            area_m2=area_m2,
+            area_m2_nonrehab=area_m2_nonrehab,
+            area_m2_rehab=area_m2_rehab,
+            change_energy_MWh=change_energy_MWh,
+            change_energy_pct=change_energy_pct,
+            cost_wage=cost_wage,
+            demand_biomass=demand_biomass,
+            demand_ediesel=demand_ediesel,
+            demand_electricity=demand_electricity,
+            demand_emethan=demand_emethan,
+            demand_emplo=demand_emplo,
+            demand_emplo_new=demand_emplo_new,
+            demand_epetrol=demand_epetrol,
+            demand_heat_nonrehab=demand_heat_nonrehab,
+            demand_heat_rehab=demand_heat_rehab,
+            demand_heatpump=demand_heatpump,
+            emplo_existing=emplo_existing,
+            energy=energy,
+            fec_factor_averaged=fec_factor_averaged,
+            invest=invest,
+            invest_pa=invest_pa,
+            invest_per_x=invest_per_x,
+            pct_nonrehab=pct_nonrehab,
+            pct_of_wage=pct_of_wage,
+            pct_rehab=pct_rehab,
+            rate_rehab_pa=rate_rehab_pa,
+            ratio_wage_to_emplo=ratio_wage_to_emplo,
+        )
+
+
+@dataclass
+class CO2eChangePOperationElecElcon:
+    change_energy_MWh: float = None  # type: ignore
+    change_energy_pct: float = None  # type: ignore
+    demand_biomass: float = None  # type: ignore
+    demand_change: float = None  # type: ignore
+    demand_ediesel: float = None  # type: ignore
+    demand_electricity: float = None  # type: ignore
+    demand_emethan: float = None  # type: ignore
+    demand_heatpump: float = None  # type: ignore
+    energy: float = None  # type: ignore
+
+    @classmethod
+    def calc(
+        cls,
+        inputs: Inputs,
+        what: str,
+        a18: A18,
+    ) -> "CO2eChangePOperationElecElcon":
+
+        demand_biomass = 0
+        demand_heatpump = 0
+        demand_ediesel = 0
+        demand_emethan = 0
+
+        demand_change = inputs.ass("Ass_B_D_fec_elec_elcon_change")
+        energy = getattr(a18, what).energy * (1 + demand_change)
+
+        demand_electricity = energy
+
+        change_energy_MWh = energy - getattr(a18, what).energy
+        change_energy_pct = div(change_energy_MWh, getattr(a18, what).energy)
+
+        return cls(
+            change_energy_MWh=change_energy_MWh,
+            change_energy_pct=change_energy_pct,
+            demand_biomass=demand_biomass,
+            demand_change=demand_change,
+            demand_ediesel=demand_ediesel,
+            demand_electricity=demand_electricity,
+            demand_emethan=demand_emethan,
+            demand_heatpump=demand_heatpump,
+            energy=energy,
+        )
+
+
+@dataclass
+class CO2eChangePOperationElecHeatpump:
+    change_energy_MWh: float = None  # type: ignore
+    demand_electricity: float = None  # type: ignore
+    energy: float = None  # type: ignore
+
+    @classmethod
+    def calc(
+        cls, inputs: Inputs, what: str, a18: A18, p_operation_heat: Any
+    ) -> "CO2eChangePOperationElecHeatpump":
+
+        energy = p_operation_heat.demand_heatpump / inputs.fact(
+            "Fact_R_S_heatpump_mean_annual_performance_factor_all"
+        )
+        demand_electricity = energy
+        change_energy_MWh = energy - getattr(a18, what).energy
+
+        return cls(
+            change_energy_MWh=change_energy_MWh,
+            demand_electricity=demand_electricity,
+            energy=energy,
+        )
+
+
+@dataclass
+class CO2eChangePOperationVehicles:
+    change_energy_MWh: float = None  # type: ignore
+    change_energy_pct: float = None  # type: ignore
+    demand_biomass: float = None  # type: ignore
+    demand_change: float = None  # type: ignore
+    demand_ediesel: float = None  # type: ignore
+    demand_electricity: float = None  # type: ignore
+    demand_emethan: float = None  # type: ignore
+    demand_epetrol: float = None  # type: ignore
+    demand_heatpump: float = None  # type: ignore
+    energy: float = None  # type: ignore
+
+    @classmethod
+    def calc(
+        cls,
+        inputs: Inputs,
+        what: str,
+        a18: A18,
+    ) -> "CO2eChangePOperationVehicles":
+
+        demand_electricity = 0
+        demand_biomass = 0
+        demand_heatpump = 0
+        demand_emethan = 0
+
+        demand_change = inputs.ass("Ass_B_D_fec_vehicles_change")
+        energy = getattr(a18, what).energy * (1 + demand_change)
+
+        demand_epetrol = div(
+            energy * a18.s_petrol.energy,
+            a18.s_petrol.energy + a18.s_diesel.energy,
+        )
+        demand_ediesel = div(
+            energy * a18.s_diesel.energy,
+            a18.s_petrol.energy + a18.s_diesel.energy,
+        )
+
+        change_energy_MWh = energy - getattr(a18, what).energy
+        change_energy_pct = div(change_energy_MWh, getattr(a18, what).energy)
+
+        return cls(
+            change_energy_MWh=change_energy_MWh,
+            change_energy_pct=change_energy_pct,
+            demand_biomass=demand_biomass,
+            demand_change=demand_change,
+            demand_ediesel=demand_ediesel,
+            demand_electricity=demand_electricity,
+            demand_emethan=demand_emethan,
+            demand_epetrol=demand_epetrol,
+            demand_heatpump=demand_heatpump,
+            energy=energy,
+        )
